requirements.txt
streamlit
pandas
requests
python-dotenv
tenacity
langchain
langchain-community
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
langchain-ollama


